[TC] Regenerated demo for tc_tc_capture_mode_sam_9x60_ek_iar
MH3-66985

+++ apps/tc/tc_capture_mode/firmware/src/config/sam_9x60_ek_iar/peripheral/pwm/plib_pwm.c
@@ -77,17 +77,17 @@ void PWM_Initialize (void)
 
 void PWM_ChannelsStart (PWM_CHANNEL_MASK channelMask)
 {
-    PWM_REGS->PWM_ENA = channelMask;
+    PWM_REGS->PWM_ENA = (uint32_t)channelMask;
 }
 
 void PWM_ChannelsStop (PWM_CHANNEL_MASK channelMask)
 {
-    PWM_REGS->PWM_DIS = channelMask;
+    PWM_REGS->PWM_DIS = (uint32_t)channelMask;
 }
 
 void PWM_ChannelPeriodSet (PWM_CHANNEL_NUM channel, uint32_t period)
 {
-    if(0 != ((PWM_REGS->PWM_SR >> channel) & 0x1U))
+    if(0U != ((PWM_REGS->PWM_SR >> channel) & 0x1U))
     {
         PWM_REGS->PWM_CH_NUM[channel].PWM_CMR |= PWM_CMR_CPD_Msk;
         PWM_REGS->PWM_CH_NUM[channel].PWM_CUPD = period;
@@ -105,7 +105,7 @@ uint32_t PWM_ChannelPeriodGet (PWM_CHANNEL_NUM channel)
 
 void PWM_ChannelDutySet(PWM_CHANNEL_NUM channel, uint32_t duty)
 {
-    if(0 != ((PWM_REGS->PWM_SR >> channel) & 0x1U))
+    if(0U != ((PWM_REGS->PWM_SR >> channel) & 0x1U))
     {
         PWM_REGS->PWM_CH_NUM[channel].PWM_CMR &= ~PWM_CMR_CPD_Msk;
         PWM_REGS->PWM_CH_NUM[channel].PWM_CUPD = duty;
@@ -118,12 +118,12 @@ void PWM_ChannelDutySet(PWM_CHANNEL_NUM channel, uint32_t duty)
 
 void PWM_ChannelCounterEventEnable (PWM_CHANNEL_MASK channelMask)
 {
-    PWM_REGS->PWM_IER = channelMask;
+    PWM_REGS->PWM_IER = (uint32_t)channelMask;
 }
 
 void PWM_ChannelCounterEventDisable (PWM_CHANNEL_MASK channelMask)
 {
-    PWM_REGS->PWM_IDR = channelMask;
+    PWM_REGS->PWM_IDR = (uint32_t)channelMask;
 }
